+++ tests/testthat/test-interpolate.conc.R
@@ -359,10 +359,91 @@ test_that("interp.extrap.conc", {
 })
 
 test_that("interp.extrap.conc.dose", {
-  expect_equal(
-    interp.extrap.conc.dose(conc=c(0, 1, 2, 1, 0.5, 0.25),
-                            time=c(-1, 1:6),
-                            time.dose=0,
-                            time.out=c(-1, -0.1, 0, 0.1, 7)),
-    c(0, 0, 0, 0.1, 0.125))
+  # Check inputs
+  expect_error(interp.extrap.conc.dose(conc=c(0, 1, 2, 1, 0.5, 0.25),
+                                       time=c(-1, 1:5),
+                                       time.dose=0, route="foo", duration.dose=NA,
+                                       time.out=c(-1, -0.1, 0, 0.1, 7), out.after=FALSE),
+               regexp="route.dose must be either 'extravascular' or 'intravascular'",
+               info="Route must be valid")
+  expect_error(interp.extrap.conc.dose(conc=c(0, 1, 2, 1, 0.5, 0.25),
+                                       time=c(-1, 1:5),
+                                       time.dose=0, route=c("extravascular", "extravascular"), duration.dose=NA,
+                                       time.out=c(-1, -0.1, 0, 0.1, 7), out.after=FALSE),
+               regexp="route.dose must either be a scalar or the same length as time.dose",
+               info="Route must have the correct length")
+  expect_error(interp.extrap.conc.dose(conc=c(0, 1, 2, 1, 0.5, 0.25),
+                                       time=c(-1, 1:5),
+                                       time.dose=0, route="extravascular", duration.dose="A",
+                                       time.out=c(-1, -0.1, 0, 0.1, 7), out.after=FALSE),
+               regexp="duration.dose must be NA or a number.",
+               info="duration.dose must be NA or a number (character).")
+  expect_error(interp.extrap.conc.dose(conc=c(0, 1, 2, 1, 0.5, 0.25),
+                                       time=c(-1, 1:5),
+                                       time.dose=0, route="extravascular", duration.dose=factor("A"),
+                                       time.out=c(-1, -0.1, 0, 0.1, 7), out.after=FALSE),
+               regexp="duration.dose must be NA or a number.",
+               info="duration.dose must be NA or a number (factor).")
+  expect_error(interp.extrap.conc.dose(conc=c(0, 1, 2, 1, 0.5, 0.25),
+                                       time=c(-1, 1:5),
+                                       time.dose=0, route="extravascular", duration.dose=c(1, NA),
+                                       time.out=c(-1, -0.1, 0, 0.1, 7), out.after=FALSE),
+               regexp="duration.dose must either be a scalar or the same length as time.dose",
+               info="duration.dose must match the length of time.dose or be a scalar.")
+
+  expect_equal(interp.extrap.conc.dose(conc=c(0, 1, 2, 1, 0.5, 0.25),
+                                       time=c(-1, 1:5),
+                                       time.dose=0,
+                                       time.out=-2),
+               structure(0, Method="Before all events"),
+               info="Interpolation before all events yields conc.origin which defaults to zero.")
+
+  expect_equal(interp.extrap.conc.dose(conc=c(0, 1, 2, 1, 0.5, 0.25),
+                                       time=c(-1, 1:5),
+                                       time.dose=0,
+                                       conc.origin=NA,
+                                       time.out=-2),
+               structure(NA, Method="Before all events"),
+               info="Interpolation before all events yields conc.origin respecting its input.")
+
+  expect_equal(interp.extrap.conc.dose(conc=c(0, 1, 2, 1, 0.5, 0.25),
+                                       time=c(-1, 1:5),
+                                       time.dose=0,
+                                       time.out=-1),
+               structure(0, Method="Copy"),
+               info="When there is a concentration measurement at a time point, it is returned.")
+  
+  expect_equal(interp.extrap.conc.dose(conc=c(0, 1, 2, 1, 0.5, 0.25),
+                                       time=c(-1, 1:5),
+                                       time.dose=0,
+                                       time.out=-0.1),
+               structure(0, Method="After a concentration with either a dose or nothing at the current time"),
+               info="When the previous measurement is zero and there is no dose between, it is returned.")
+
+  expect_equal(interp.extrap.conc.dose(conc=c(0, 1, 2, 1, 0.5, 0.25),
+                                       time=c(-1, 1:5),
+                                       time.dose=0,
+                                       time.out=0),
+               structure(0, Method="After a concentration with either a dose or nothing at the current time"),
+               info="When the previous measurement is zero it is at the time of the dose, zero is returned.")
+
+  expect_equal(interp.extrap.conc.dose(conc=c(0, 1, 2, 1, 0.5, 0.25),
+                                       time=c(-1, 1:5),
+                                       time.dose=0,
+                                       time.out=0.1),
+               structure(0.1, Method="After a dose with no event and a concentration after"),
+               info="Extrapolation to a dose then interpolation between the dose and the next time works.")
+
+  expect_equal(interp.extrap.conc.dose(conc=c(0, 1, 2, 1, 0.5, 0.25),
+                                       time=c(-1, 1:5),
+                                       time.dose=0,
+                                       time.out=5),
+               structure(0.25, Method="Copy"),
+               info="Copy from after the dose.")
+
+  expect_equal(interp.extrap.conc.dose(conc=c(0, 1, 2, 1, 0.5, 0.25),
+                                       time=c(-1, 1:5),
+                                       time.dose=0,
+                                       time.out=7),
+               structure(0.0625, Method="After a concentration with either a dose or nothing at the current time"))
 })